Add hashing to environment (#94)
Why:

Hashing an environment makes it possible to verify if two initial states
are identical. This is necessary if we want to create groups of initial
states that are mutually exclusive.

How:

Compute the hash of the grid encoding with the initial agent position and
direction.

Note:

I'm not sure if the RNG state may eventually cause divergences in the
episodes and whether that should be accounted for in the hashing. It's
difficult to measure however the divergence without playing, and taking
into account the RNG state will likely give different hashing for
practically identical states. That's why my feeling is the RNG state should not
be part of the hashing.